feat: moved hook to new React package
