+++ packages/nextjs/src/index.tsx
@@ -1,0 +1,2 @@
+export * from './context';
+export * from '@wundergraph/react';
