+++ packages/react/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@wundergraph/nextjs",
-  "version": "0.1.0",
+  "name": "@wundergraph/react",
+  "version": "0.0.1",
   "license": "Apache-2.0",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
@@ -31,9 +31,9 @@
     "@types/node": "^17.0.32",
     "@types/object-hash": "^1.3.4",
     "@types/react": "^18.0.15",
-    "@wundergraph/sdk": "workspace:^0.97.0",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
+    "@wundergraph/sdk": "workspace:^0.96.1",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
     "typescript": "^4.7.2"
   },
   "peerDependencies": {
@@ -43,11 +43,8 @@
   },
   "dependencies": {
     "@wundergraph/protobuf": "workspace:^0.91.1",
-    "@wundergraph/react": "workspace:^0.0.1",
     "handlebars": "^4.7.7",
-    "next": "^12.1.6",
-    "object-hash": "^2.2.0",
-    "react-ssr-prepass": "^1.5.0"
+    "object-hash": "^2.2.0"
   },
   "sideEffects": false
 }
