+++ packages/react/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "target": "ES2019",
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "declarationDir": "./dist",
+    "jsx": "react",
+    "outDir": "./dist",
+    "typeRoots": ["./node_modules/@types", "./src/typedefinitions"]
+  },
+  "exclude": ["node_modules", "dist"]
+}
